Huge refactor.
GOTUser had a bunch of data that only really applied to the logged in user.
So created the GOTActiveUser class which decorates a GOTUser instance to store
data which we only really need about the active user (like facebook_id, email,
token, pending_email).
Also changed it so that when we're updating a user (GOTActiveUser) we don't
change the values until the web request returns.  Because of that, we need
to pass any data we want to change as a dictionary of params.
Now that we have this special GOTActiveUser object, we can archive it
and unarchive it to see who is logged in.  This removes the need to
store the FacebookID in the global settings.
We also made some changes to how the web server processed ids (they are now
more explicit) so that required changing some things in the client.